requirements.txt
# for all
numpy>=1.24.2
wandb>=0.13.10
matplotlib>=3.7.0
scipy>=1.10.0
pytorch-lightning>=2.0.3
torch-geometric>=2.2.0
shapely>=2.0.1
networkx>=3.0
Cartopy>=0.22.0
pyproj>=3.4.1
tueplots>=0.0.8
plotly>=5.15.0
copernicusmarine>=1.2.0
momo-opt>=0.1.0
cdsapi>=0.7.0
ecmwf-opendata>=0.3.8
# for dev
codespell>=2.0.0
black>=21.9b0
isort>=5.9.3
flake8>=4.0.1
pylint>=3.0.3
pre-commit>=2.15.0
scikit-image>=0.25.1
git+https://github.com/google-research/weatherbenchX.git
netCDF4==1.7.2